fixup! ISSUE-1038 Capability object for JMAP PublicAsset extension
