+++ tmail-backend/jmap/extensions/src/main/scala/com/linagora/tmail/james/jmap/publicAsset/PublicAssetsModule.scala
@@ -1,0 +1,14 @@
+package com.linagora.tmail.james.jmap.publicAsset
+
+import com.google.inject.AbstractModule
+import com.google.inject.multibindings.Multibinder
+import com.linagora.tmail.james.jmap.method.PublicAssetsCapabilityFactory
+import org.apache.james.jmap.core.CapabilityFactory
+
+class PublicAssetsModule extends AbstractModule {
+  override def configure(): Unit = {
+    Multibinder.newSetBinder(binder(), classOf[CapabilityFactory])
+      .addBinding()
+      .to(classOf[PublicAssetsCapabilityFactory])
+  }
+}
